ops/requirements.txt
PyYAML>=6.0.1
Jinja2>=3.1.2
python-dotenv>=1.0.0
kubernetes>=28.1.0
